Openstack scope refactoring
