+++ openstack/datadog_checks/openstack/settings.py
@@ -1,0 +1,12 @@
+# (C) Datadog, Inc. 2010-2017
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+
+UNSCOPED_AUTH = 'unscoped'
+
+V21_NOVA_API_VERSION = 'v2.1'
+
+DEFAULT_KEYSTONE_API_VERSION = 'v3'
+DEFAULT_NOVA_API_VERSION = V21_NOVA_API_VERSION
+FALLBACK_NOVA_API_VERSION = 'v2'
+DEFAULT_API_REQUEST_TIMEOUT = 10  # seconds
