+++ openstack/tests/test_openstack.py
@@ -14,15 +14,10 @@ import pytest
 # project
 import common
 
-from datadog_checks.openstack.openstack import (
-    OpenStackCheck,
-    OpenStackProjectScope,
-    OpenStackUnscoped,
-    KeystoneCatalog,
-    IncompleteConfig,
-    IncompleteAuthScope,
-    IncompleteIdentity
-)
+from datadog_checks.openstack.exceptions import (IncompleteConfig, IncompleteAuthScope, IncompleteIdentity)
+from datadog_checks.openstack.catalog import ServiceCatalog
+from datadog_checks.openstack.openstack import OpenStackCheck
+from datadog_checks.openstack.scopes import (OpenStackProjectScope, OpenStackUnscoped)
 
 from datadog_checks.checks import AgentCheck
 
@@ -56,12 +51,12 @@ MOCK_HTTP_PROJECTS_RESPONSE = MockHTTPResponse(response_dict=common.EXAMPLE_PROJ
 
 def _test_bad_auth_scope(scope):
     with pytest.raises(IncompleteAuthScope):
-        OpenStackProjectScope.get_auth_scope(scope)
+        OpenStackProjectScope._get_auth_scope(scope)
 
 
 def _test_bad_user(user):
     with pytest.raises(IncompleteIdentity):
-        OpenStackProjectScope.get_user_identity(user)
+        OpenStackProjectScope._get_user_identity(user)
 
 
 def test_get_auth_scope():
@@ -69,14 +64,14 @@ def test_get_auth_scope():
         _test_bad_auth_scope(scope)
 
     for scope in common.GOOD_UNSCOPED_AUTH_SCOPES:
-        auth_scope = OpenStackProjectScope.get_auth_scope(scope)
+        auth_scope = OpenStackProjectScope._get_auth_scope(scope)
         assert auth_scope is None
-        auth_scope = OpenStackUnscoped.get_auth_scope(scope)
+        auth_scope = OpenStackUnscoped._get_auth_scope(scope)
 
         assert auth_scope is None
 
     for scope in common.GOOD_AUTH_SCOPES:
-        auth_scope = OpenStackProjectScope.get_auth_scope(scope)
+        auth_scope = OpenStackProjectScope._get_auth_scope(scope)
 
         # Should pass through unchanged
         assert auth_scope == scope.get('auth_scope')
@@ -87,7 +82,7 @@ def test_get_user_identity():
         _test_bad_user(user)
 
     for user in common.GOOD_USERS:
-        parsed_user = OpenStackProjectScope.get_user_identity(user)
+        parsed_user = OpenStackProjectScope._get_user_identity(user)
         assert parsed_user == {'methods': ['password'], 'password': user}
 
 
@@ -102,7 +97,7 @@ def test_from_config():
         OpenStackProjectScope.from_config(init_config, bad_instance_config)
 
     with mock.patch(
-        'datadog_checks.openstack.openstack.OpenStackProjectScope.request_auth_token',
+        'datadog_checks.openstack.openstack.OpenStackProjectScope._request_auth_token',
         return_value=MOCK_HTTP_RESPONSE
     ):
         append_config = good_instance_config.copy()
@@ -128,15 +123,15 @@ def test_unscoped_from_config():
     mock_http_response['token'].pop('project')
     mock_response = MockHTTPResponse(response_dict=mock_http_response, headers={'X-Subject-Token': 'fake_token'})
     with mock.patch(
-        'datadog_checks.openstack.openstack.OpenStackUnscoped.request_auth_token',
+        'datadog_checks.openstack.openstack.OpenStackUnscoped._request_auth_token',
         return_value=mock_response
     ):
         with mock.patch(
-            'datadog_checks.openstack.openstack.OpenStackUnscoped.request_project_list',
+            'datadog_checks.openstack.openstack.OpenStackUnscoped._request_project_list',
             return_value=MOCK_HTTP_PROJECTS_RESPONSE
         ):
             with mock.patch(
-                'datadog_checks.openstack.openstack.OpenStackUnscoped.get_token_for_project',
+                'datadog_checks.openstack.openstack.OpenStackUnscoped._get_token_for_project',
                 return_value=MOCK_HTTP_RESPONSE
             ):
                 append_config = good_instance_config.copy()
@@ -153,20 +148,20 @@ def test_unscoped_from_config():
 
 
 def test_get_nova_endpoint():
-    assert KeystoneCatalog.get_nova_endpoint(
+    assert ServiceCatalog._get_nova_endpoint(
         common.EXAMPLE_AUTH_RESPONSE) == u'http://10.0.2.15:8774/v2.1/0850707581fe4d738221a72db0182876'
-    assert KeystoneCatalog.get_nova_endpoint(
+    assert ServiceCatalog._get_nova_endpoint(
         common.EXAMPLE_AUTH_RESPONSE,
         nova_api_version='v2') == u'http://10.0.2.15:8773/'
 
 
 def test_get_neutron_endpoint():
-    assert KeystoneCatalog.get_neutron_endpoint(common.EXAMPLE_AUTH_RESPONSE) == u'http://10.0.2.15:9292'
+    assert ServiceCatalog._get_neutron_endpoint(common.EXAMPLE_AUTH_RESPONSE) == u'http://10.0.2.15:9292'
 
 
 def test_from_auth_response():
-    catalog = KeystoneCatalog.from_auth_response(common.EXAMPLE_AUTH_RESPONSE, 'v2.1')
-    assert isinstance(catalog, KeystoneCatalog)
+    catalog = ServiceCatalog.from_auth_response(common.EXAMPLE_AUTH_RESPONSE, 'v2.1')
+    assert isinstance(catalog, ServiceCatalog)
     assert catalog.neutron_endpoint == u'http://10.0.2.15:9292'
     assert catalog.nova_endpoint == u'http://10.0.2.15:8774/v2.1/0850707581fe4d738221a72db0182876'
 
@@ -179,7 +174,7 @@ def test_ensure_auth_scope(aggregator):
         openstack_check.get_scope_for_instance(instance)
 
     with mock.patch(
-        'datadog_checks.openstack.openstack.OpenStackProjectScope.request_auth_token',
+        'datadog_checks.openstack.openstack.OpenStackProjectScope._request_auth_token',
         return_value=MOCK_HTTP_RESPONSE
     ):
         scope = openstack_check.ensure_auth_scope(instance)
